Refactor Myna into Experiment / Suggestion. Test rewarding.
Instead of a single Myna object, mirroring the REST API, I've split
the code into an Experiment and a Suggestion object. This makes the
API easier to understand, as we don't have to worry about copying
tokens etc. around -- a Suggestion knows the experiment it is
associated witha and also holds its token.

The old basic code is still around as I haven't moved the cookie
functionality out.